requirements.txt
docling
openai
supabase
python-dotenv
termcolor
tiktoken
transformers
